+++ lib/src/models/ble_scan_result.dart
@@ -4,6 +4,8 @@ class BleDevice {
   String deviceId;
   String? name;
   bool? isPaired;
+  bool
+      isConnected; // TODO: make this read only and perhaps create a getter which returns a real time value
   Uint8List? manufacturerDataHead;
   Uint8List? manufacturerData;
   int? rssi;
@@ -14,6 +16,7 @@ class BleDevice {
     required this.deviceId,
     this.rssi,
     this.isPaired,
+    this.isConnected = false,
     Uint8List? manufacturerData,
     Uint8List? manufacturerDataHead,
     this.services = const [],